+++ lib/env.dart
@@ -1,0 +1,5 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+class Env {
+  Env._();
+}